Refactor tamonomies into theme
