+++ wp-content/themes/clarity/inc/admin/agency_taxonomies/taxonomies/taxonomy.php
